Make sure first_byte_timeout is applied on reused backend connection
This avoids getting stuck with a stalled backend connection until
reaching backend_idle_timeout.

We now wait for the waiter to release the fd for a limited amount of
time (first_byte_timeout). In case we reach the timeout, the backend
connection is killed to make sure we are not going to reuse it.

Fixes #1772.